+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Next
+- Add playback history to queue panel
 - Update dependencies
 - Drop support for Windows 7 and 8
 - Drop support for macOS 10.13 and 10.14